CL: - Various configuration changes to build the project correctly
