--- lib/shadcn-plugin.ts
@@ -1,195 +1,0 @@
-import colors from "tailwindcss/colors"
-import { fontFamily } from "tailwindcss/defaultTheme"
-import plugin from "tailwindcss/plugin"
-
-export const shadcnPlugin = plugin(
-  function ({ addBase }) {
-    // colors
-    addBase({
-      ":root": {
-        "--background": "240, 50%, 95%",
-        "--foreground": "0, 0%, 13%",
-        "--muted": "210 40% 96.1%",
-        "--muted-foreground": "215.4 16.3% 46.9%",
-        "--popover": "0 0% 100%",
-        "--popover-foreground": "222.2 47.4% 11.2%",
-        "--border": "214.3 31.8% 91.4%",
-        "--input": "214.3 31.8% 91.4%",
-        "--card": "0 0% 100%",
-        "--card-foreground": "222.2 47.4% 11.2%",
-        "--primary": "222.2 88% 11.2%",
-        "--primary-foreground": "175 42% 88%",
-        "--secondary": "210 40% 96.1%",
-        "--secondary-foreground": "17 10% 59%",
-        "--accent": "210 40% 96.1%",
-        "--accent-foreground": "60 16% 91%",
-        "--destructive": "0 100% 50%",
-        "--destructive-foreground": "210 40% 98%",
-        "--ring": "215 20.2% 65.1%",
-        "--radius": "0.5rem",
-      },
-      "[data-theme='dark']": {
-        "--background": "224 71% 4%",
-        "--foreground": "213 31% 91%",
-        "--muted": "223 47% 11%",
-        "--muted-foreground": "215.4 16.3% 56.9%",
-        "--accent": "216 34% 17%",
-        "--accent-foreground": "210 40% 98%",
-        "--popover": "224 71% 4%",
-        "--popover-foreground": "215 20.2% 65.1%",
-        "--border": "216 34% 17%",
-        "--input": "216 34% 17%",
-        "--card": "224 71% 4%",
-        "--card-foreground": "213 31% 91%",
-        "--primary": "210 40% 98%",
-        "--primary-foreground": "222.2 47.4% 1.2%",
-        "--secondary": "222.2 47.4% 11.2%",
-        "--secondary-foreground": "210 40% 98%",
-        "--destructive": "0 63% 31%",
-        "--destructive-foreground": "210 40% 98%",
-        "--ring": "216 34% 17%",
-        "--radius": "0.5rem",
-      },
-    })
-    // classes
-    addBase({
-      "*": {
-        "@layer border-border": {},
-      },
-      body: {
-        "@apply bg-background text-foreground": {},
-        "font-feature-settings": '"rlig" 1, "calt" 1',
-      },
-      ".animation-delay-2000": {
-        "animation-delay": "2s",
-      },
-      ".animation-delay-4000": {
-        "animation-delay": "4s",
-      },
-    })
-  },
-  {
-    theme: {
-      container: {
-        center: true,
-        padding: "2rem",
-        screens: {
-          "2xl": "1400px",
-        },
-      },
-      extend: {
-        colors: {
-          background: "hsl(var(--background))",
-          foreground: "hsl(var(--foreground))",
-          primary: {
-            DEFAULT: "hsl(var(--primary))",
-            foreground: "hsl(var(--primary-foreground))",
-          },
-          secondary: {
-            DEFAULT: "hsl(var(--secondary))",
-            foreground: "hsl(var(--secondary-foreground))",
-          },
-          destructive: {
-            DEFAULT: "hsl(var(--destructive))",
-            foreground: "hsl(var(--destructive-foreground))",
-          },
-          muted: {
-            DEFAULT: "hsl(var(--muted))",
-            foreground: "hsl(var(--muted-foreground))",
-          },
-          accent: {
-            DEFAULT: "hsl(var(--accent))",
-            foreground: "hsl(var(--accent-foreground))",
-          },
-          popover: {
-            DEFAULT: "hsl(var(--popover))",
-            foreground: "hsl(var(--popover-foreground))",
-          },
-          card: {
-            DEFAULT: "hsl(var(--card))",
-            foreground: "hsl(var(--card-foreground))",
-          },
-
-          neutral: {
-            50: "#fafafa",
-            100: "#f4f4f4",
-            200: "#e6e6e6",
-            300: "#d4d4d4",
-            400: "#a2a2a2",
-            500: "#727272",
-            600: "#535353",
-            700: "#404040",
-            800: "#272727",
-            900: "#181818",
-          },
-          indigo: colors.indigo,
-          slate: colors.slate,
-          blue: {
-            DEFAULT: "#81c7da",
-            50: "#effafe",
-            100: "#dbeef5",
-            200: "#b5dee9",
-            300: "#81c7da",
-            400: "#46b6d4",
-            500: "#14a2ca",
-            600: "#0084ad",
-            700: "#006b8d",
-            800: "#005b77",
-            900: "#004d65",
-          },
-          red: {
-            DEFAULT: "#ff6d62",
-            50: "#fef2f1",
-            100: "#ffe2e0",
-            200: "#ffcbc7",
-            300: "#ffa59e",
-            400: "#ff6d62",
-            500: "#f53d2b",
-            600: "#dd2700",
-            700: "#b92000",
-            800: "#9a1b07",
-            900: "#7f1e13",
-          },
-        },
-        borderRadius: {
-          lg: `var(--radius)`,
-          md: `calc(var(--radius) - 2px)`,
-          sm: "calc(var(--radius) - 4px)",
-        },
-        fontFamily: {
-          sans: ["var(--font-sans)", ...fontFamily.sans],
-          heading: ["var(--font-heading)", ...fontFamily.sans],
-        },
-        keyframes: {
-          "accordion-down": {
-            from: { height: "0" },
-            to: { height: "var(--radix-accordion-content-height)" },
-          },
-          "accordion-up": {
-            from: { height: "var(--radix-accordion-content-height)" },
-            to: { height: "0" },
-          },
-          blob: {
-            "0%": {
-              transform: "translate(0px, 0px) scale(1)",
-            },
-            "33%": {
-              transform: "translate(30px, -50px) scale(1.1)",
-            },
-            "66%": {
-              transform: "translate(-20px, 20px) scale(0.9)",
-            },
-            "100%": {
-              transform: "translate(0px, 0px) scale(1)",
-            },
-          },
-        },
-        animation: {
-          "accordion-down": "accordion-down 0.2s ease-out",
-          "accordion-up": "accordion-up 0.2s ease-out",
-          blob: "blob 8s infinite",
-        },
-      },
-    },
-  }
-)
